change UI in all forms
