+++ src/app/loginpage/page.tsx
@@ -10,7 +10,10 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
+import { useRouter } from "next/navigation";
+
 import { zodResolver } from "@hookform/resolvers/zod";
+import Image from "next/image";
 import Link from "next/link";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
@@ -28,16 +31,35 @@ export default function LoginForm() {
       password: "",
     },
   });
+  const router = useRouter();
 
   const onSubmit = (data: unknown) => {
     console.log("Login Data:", data);
     alert("Login successful");
+    router.push("/dashboard");
   };
 
   return (
-    <div className="flex min-h-screen items-center justify-center bg-gray-100 px-4">
-      <div className="w-full max-w-md rounded-lg bg-white p-6 shadow-lg">
-        <h2 className="mb-4 text-center font-semibold text-2xl">Login</h2>
+    <div className="flex min-h-screen">
+      {/* Left Section */}
+      <div className="relative hidden w-1/2 lg:flex">
+        <Image
+          src="/image.png"
+          alt="Background Illustration"
+          layout="fill"
+          objectFit="cover"
+          className="absolute"
+        />
+      </div>
+
+      {/* Right Section - Form */}
+      <div className="flex w-full flex-col items-center justify-center bg-gray-900 p-10 text-white shadow-lg lg:w-1/2">
+        <h2 className="mb-6 text-center font-bold text-3xl text-teal-400">
+          Login Form
+        </h2>
+        <p className="mb-6 text-center text-gray-300">
+          Welcome to whether Application!
+        </p>
         <Form {...form}>
           <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
             <FormField
@@ -64,11 +86,7 @@ export default function LoginForm() {
                 <FormItem>
                   <FormLabel>Password</FormLabel>
                   <FormControl>
-                    <Input
-                      type="password"
-                      placeholder="Enter your password"
-                      {...field}
-                    />
+                    <Input type="password" {...field} />
                   </FormControl>
                   <FormMessage />
                 </FormItem>
@@ -82,17 +100,17 @@ export default function LoginForm() {
                 Forgot Password?
               </Link>
             </div>
-            <Button type="submit" className="w-full">
+            <Button
+              type="submit"
+              className="w-full bg-sky-700 hover:bg-blue-700"
+            >
               Login
             </Button>
           </form>
         </Form>
         <p className="mt-4 text-center text-sm">
           Don't have an account?{" "}
-          <Link
-            href="/registrationform"
-            className="text-blue-600 hover:underline"
-          >
+          <Link href="/registration" className="text-blue-600 hover:underline">
             Register here
           </Link>
         </p>
