+++ src/app/validationpass/page.tsx
@@ -10,6 +10,7 @@ import {
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { zodResolver } from "@hookform/resolvers/zod";
+import Image from "next/image";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
@@ -57,62 +58,82 @@ export default function ForgotPasswordForm() {
   };
 
   return (
-    <Form {...form}>
-      <form
-        onSubmit={form.handleSubmit(onSubmit)}
-        className="mx-auto mt-60 max-w-md space-y-4 rounded-lg border p-4 shadow-md "
-      >
-        {/* Email Input */}
-        <FormField
-          control={form.control}
-          name="email"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>
-                Email <span className="text-red-500">*</span>
-              </FormLabel>
-              <FormControl>
-                <Input type="email" {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
+    <div className="flex min-h-screen">
+      {/* Left Section */}
+      <div className="relative hidden w-1/2 lg:flex">
+        <Image
+          src="/image.png"
+          alt="Background Illustration"
+          layout="fill"
+          objectFit="cover"
+          className="absolute"
         />
+      </div>
 
-        {!emailSent ? (
-          <Button type="button" onClick={handleEmailSubmit} className="w-full">
-            Send OTP
-          </Button>
-        ) : (
-          <>
-            {/* OTP Input */}
+      {/* Right Section - Form */}
+      <div className=" flex-col items-center justify-center bg-gray-900 p-10 text-white shadow-lg lg:w-1/2">
+        <Form {...form}>
+          <form
+            onSubmit={form.handleSubmit(onSubmit)}
+            className="mx-auto mt-60 max-w-md space-y-4 rounded-lg border p-4 shadow-md "
+          >
+            {/* Email Input */}
             <FormField
               control={form.control}
-              name="otp"
+              name="email"
               render={({ field }) => (
                 <FormItem>
-                  <FormLabel>Enter OTP</FormLabel>
+                  <FormLabel>
+                    Email <span className="text-red-500">*</span>
+                  </FormLabel>
                   <FormControl>
-                    <Input
-                      placeholder="Enter OTP sent to your email"
-                      {...field}
-                    />
+                    <Input type="email" {...field} />
                   </FormControl>
                   <FormMessage />
                 </FormItem>
               )}
             />
 
-            <Button
-              type="submit"
-              className="w-full"
-              disabled={!form.watch("otp")}
-            >
-              Submit
-            </Button>
-          </>
-        )}
-      </form>
-    </Form>
+            {!emailSent ? (
+              <Button
+                type="button"
+                onClick={handleEmailSubmit}
+                className="item-center flex bg-blue-500 hover:bg-sky-700"
+              >
+                Send OTP
+              </Button>
+            ) : (
+              <>
+                {/* OTP Input */}
+                <FormField
+                  control={form.control}
+                  name="otp"
+                  render={({ field }) => (
+                    <FormItem>
+                      <FormLabel>Enter OTP</FormLabel>
+                      <FormControl>
+                        <Input
+                          placeholder="Enter OTP sent to your email"
+                          {...field}
+                        />
+                      </FormControl>
+                      <FormMessage />
+                    </FormItem>
+                  )}
+                />
+
+                <Button
+                  type="submit"
+                  className="w-full bg-blue-950 hover:bg-sky-700"
+                  disabled={!form.watch("otp")}
+                >
+                  Submit
+                </Button>
+              </>
+            )}
+          </form>
+        </Form>
+      </div>
+    </div>
   );
 }
