+++ src/store/index.js
@@ -1,0 +1,8 @@
+import { configureStore } from "@reduxjs/toolkit";
+import registrationReducer from "./registrationSlice";
+
+export const store = configureStore({
+  reducer: {
+    registration: registrationReducer,
+  },
+});
